Feature/tax extension

+++ CommonWallet/Classes/Common/Model/WalletAsset.swift
@@ -12,12 +12,12 @@ public struct WalletAsset {
     public let symbol: String
     public let details: LocalizableResource<String>
     public let identifier: IRAssetId
-    public let precision: UInt8
+    public let precision: Int16
 
     public init(identifier: IRAssetId,
                 symbol: String,
                 details: LocalizableResource<String>,
-                precision: UInt8) {
+                precision: Int16) {
         self.identifier = identifier
         self.symbol = symbol
         self.details = details
